Happstack.Server.Client

Description

a very simple interface for acting as an HTTP client. This is mostly used for things like Happstack.Server.Proxy. You are more likely to want a library like http-enumerator http://hackage.haskell.org/package/http-enumerator.

Documentation

getResponse :: Request -> IO (Either String Response)

Sends the serialized request to the host defined in the request and attempts to parse response upon arrival.