Happstack.Server.Internal.Monads

Description

This module defines the Monad stack used by Happstack. You mostly don't want to be looking in here. Look in Happstack.Server.Monads instead.

Documentation

type Web a = WebT IO a

An alias for WebT when using IO.

type ServerPart a = ServerPartT IO a

An alias for ServerPartT IO

newtype ServerPartT m a

ServerPartT is a rich, featureful monad for web development.

see also: simpleHTTP, ServerMonad, FilterMonad, WebMonad, and HasRqData

Constructors

ServerPartT

Fields

unServerPartT :: ReaderT Request (WebT m) a

runServerPartT :: ServerPartT m a -> Request -> WebT m a

Particularly useful when combined with runWebT to produce a m (Maybe Response) from a Request.

withRequest :: (Request -> WebT m a) -> ServerPartT m a

function for lifting WebT to ServerPartT

NOTE: This is mostly for internal use. If you want to access the Request in user-code see askRq from ServerMonad.

 do request <- askRq
    ...

anyRequest :: Monad m => WebT m a -> ServerPartT m a

A constructor for a ServerPartT when you don't care about the request.

NOTE: This is mostly for internal use. If you think you need to use it in your own code, you might consider asking on the mailing list or IRC to find out if there is an alternative solution.

mapServerPartT :: (UnWebT m a -> UnWebT n b) -> ServerPartT m a -> ServerPartT n b

Apply a function to transform the inner monad of ServerPartT m.

Often used when transforming a monad with ServerPartT, since simpleHTTP requires a ServerPartT IO a. Refer to UnWebT for an explanation of the structure of the monad.

Here is an example. Suppose you want to embed an ErrorT into your ServerPartT to enable throwError and catchError in your Monad.

 type MyServerPartT e m a = ServerPartT (ErrorT e m) a

Now suppose you want to pass MyServerPartT into a function that demands a ServerPartT IO a (e.g. simpleHTTP). You can provide the function:

   unpackErrorT :: (Monad m, Show e) => UnWebT (ErrorT e m) a -> UnWebT m a
   unpackErrorT et = do
      eitherV <- runErrorT et
      return $ case eitherV of
          Left err -> Just (Left $ toResponse $
                                   "Catastrophic failure " ++ show err
                           , filterFun $ \r -> r{rsCode = 500})
          Right x -> x

With unpackErrorT you can now call simpleHTTP. Just wrap your ServerPartT list.

  simpleHTTP nullConf $ mapServerPartT unpackErrorT (myPart `catchError` myHandler)

Or alternatively:

  simpleHTTP' unpackErrorT nullConf (myPart `catchError` myHandler)

Also see spUnwrapErrorT for a more sophisticated version of this function.

mapServerPartT' :: (Request -> UnWebT m a -> UnWebT n b) -> ServerPartT m a -> ServerPartT n b

A variant of mapServerPartT where the first argument also takes a Request. Useful if you want to runServerPartT on a different ServerPartT inside your monad (see spUnwrapErrorT).

class Monad m => ServerMonad m where

The ServerMonad class provides methods for reading or locally modifying the Request. It is essentially a specialized version of the MonadReader class. Providing the unique names, askRq and localRq makes it easier to use ServerPartT and ReaderT together.

Methods

askRq :: m Request

localRq :: (Request -> Request) -> m a -> m a

data SetAppend a

A monoid operation container. If a is a monoid, then SetAppend is a monoid with the following behaviors:

  Set    x `mappend` Append y = Set    (x `mappend` y)
  Append x `mappend` Append y = Append (x `mappend` y)
  _        `mappend` Set y    = Set y

A simple way of summarizing this is, if the right side is Append, then the right is appended to the left. If the right side is Set, then the left side is ignored.

Constructors

Set a

Append a

extract :: SetAppend t -> t

Extract the value from a SetAppend. Note that a SetAppend is actually a CoPointed from: http://hackage.haskell.org/packages/archive/category-extras/latest/doc/html/Control-Functor-Pointed.html But lets not drag in that dependency. yet...

type FilterFun a = SetAppend (Dual (Endo a))

FilterFun is a lot more fun to type than SetAppend (Dual (Endo a)).

unFilterFun :: FilterFun a -> a -> a

filterFun :: (a -> a) -> FilterFun a

turn a function into a FilterFun. Primarily used with mapServerPartT

newtype FilterT a m b

Constructors

FilterT

Fields

unFilterT :: WriterT (FilterFun a) m b

class Monad m => FilterMonad a m | m -> a where

A set of functions for manipulating filters.

ServerPartT implements FilterMonad Response so these methods are the fundamental ways of manipulating Response values.

Methods

setFilter :: (a -> a) -> m ()

Ignores all previous alterations to your filter

As an example:

 do
   composeFilter f
   setFilter g
   return "Hello World"

The setFilter g will cause the first composeFilter f to be ignored.

composeFilter :: (a -> a) -> m ()

Composes your filter function with the existing filter function.

getFilter :: m b -> m (b, a -> a)

Retrieves the filter from the environment.

ignoreFilters :: FilterMonad a m => m ()

Resets all your filters. An alias for setFilter id.

newtype WebT m a

The basic Response building object.

Constructors

WebT

Fields

unWebT :: ErrorT Response (FilterT Response (MaybeT m)) a

type UnWebT m a = m (Maybe (Either Response a, FilterFun Response))

UnWebT is almost exclusively used with mapServerPartT. If you are not using mapServerPartT then you do not need to wrap your head around this type. If you are -- the type is not as complex as it first appears.

It is worth discussing the unpacked structure of WebT a bit as it's exposed in mapServerPartT and mapWebT.

A fully unpacked WebT has a structure that looks like:

 ununWebT $ WebT m a :: m (Maybe (Either Response a, FilterFun Response))

So, ignoring m, as it is just the containing Monad, the outermost layer is a Maybe. This is Nothing if mzero was called or Just (Either Response a, SetAppend (Endo Response)) if mzero wasn't called. Inside the Maybe, there is a pair. The second element of the pair is our filter function FilterFun Response. FilterFun Response is a type alias for SetAppend (Dual (Endo Response)). This is just a wrapper for a Response -> Response function with a particular Monoid behavior. The value

  Append (Dual (Endo f))

Causes f to be composed with the previous filter.

  Set (Dual (Endo f))

Causes f to not be composed with the previous filter.

Finally, the first element of the pair is either Left Response or Right a.

Another way of looking at all these pieces is from the behaviors they control. The Maybe controls the mzero behavior. Set (Endo f) comes from the setFilter behavior. Likewise, Append (Endo f) is from composeFilter. Left Response is what you get when you call finishWith and Right a is the normal exit.

An example case statement looks like:

  ex1 webt = do
    val <- ununWebT webt
    case val of
        Nothing -> Nothing  -- this is the interior value when mzero was used
        Just (Left r, f) -> Just (Left r, f) -- r is the value that was passed into "finishWith"
                                             -- f is our filter function
        Just (Right a, f) -> Just (Right a, f) -- a is our normal monadic value
                                               -- f is still our filter function

class Monad m => WebMonad a m | m -> a where

WebMonad provides a means to end the current computation and return a Response immediately. This provides an alternate escape route. In particular it has a monadic value of any type. And unless you call setFilter id first your response filters will be applied normally.

Extremely useful when you're deep inside a monad and decide that you want to return a completely different content type, since it doesn't force you to convert all your return types to Response early just to accommodate this.

see also: escape and escape'

Methods

finishWith

Arguments

:: a
value to return (For ServerPart, a will always be the type Response)

-> m b

escape :: (WebMonad a m, FilterMonad a m) => m a -> m b

Used to ignore all your filters and immediately end the computation. A combination of ignoreFilters and finishWith.

escape' :: (WebMonad a m, FilterMonad a m) => a -> m b

An alternate form of escape that can be easily used within a do block.

ununWebT :: WebT m a -> UnWebT m a

For when you really need to unpack a WebT entirely (and not just unwrap the first layer with unWebT).

mkWebT :: UnWebT m a -> WebT m a

For wrapping a WebT back up. mkWebT . ununWebT = id

mapWebT :: (UnWebT m a -> UnWebT n b) -> WebT m a -> WebT n b

See mapServerPartT for a discussion of this function.

localContext :: Monad m => (WebT m a -> WebT m' a) -> ServerPartT m a -> ServerPartT m' a

This is kinda like a very oddly shaped mapServerPartT or mapWebT. You probably want one or the other of those.

multi :: Monad m => [ServerPartT m a] -> ServerPartT m a

Deprecated: Use msum instead

Deprecated: use msum.

debugFilter :: (MonadIO m, Show a) => ServerPartT m a -> ServerPartT m a

Deprecated: This function appears to do nothing.

What is this for, exactly? I don't understand why Show a is even in the context Deprecated: This function appears to do nothing at all. If it use it, let us know why.

outputTraceMessage :: String -> a -> a

mkFailMessage :: (FilterMonad Response m, WebMonad Response m) => String -> m b

failHtml :: String -> String

escapeString :: String -> String