Happstack.Server.Internal.TimeoutIO

Documentation

data TimeoutIO

TimeoutIO is a record which abstracts out all the network IO functions needed by the request handling loop. This allows use to use the same event loop for handle both http: and https:.

Constructors

TimeoutIO

Fields

toHandle :: Handle

toPutLazy :: ByteString -> IO ()

toPut :: ByteString -> IO ()

toGetContents :: IO ByteString

toSendFile :: FilePath -> Offset -> ByteCount -> IO ()

toShutdown :: IO ()

toSecure :: Bool